requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn==1.6.1
streamlit
Pillow
joblib